+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## unreleased
 
+* Use `doks.digitalocean.com/node-network-type` to distinguish public vs private-only nodes when retrying LB sync (CON-14209) (@pyadagiri-do)
+
 ## v0.1.69 (beta) - August 25, 2026
 
 * Retry LB sync while nodes are still cloud-provider initializing (@pyadagiri-do)